Add files via upload
The impact area is calculated based on the fact that slopes of more than 35 degrees have a risk of failure. The slope was calculated based on the DEM dataset at 0.5 meter spatial resolution, gathered by IOM drone imagery. The DEM was adjusted match a geographical point of reference, and trees and buildings removed. The additional area of susceptibility of landslide was extended by manually drawing polygons by UNHCR and ADPC, with the support of DEM topography and contour lines. These polygons are extensions of 40 degrees and above until reaching the base of the respective slope. Spatial analysis was carried out in order to provide the statistical results of the population at risk.

Risk Management Criteria and assumptions made:

• The crucial landslide trigger factor is pore pressure • The land slide failure would be sudden • When it does fail, it will have an aspect ratio of 1 to 1 • 35 degrees slope and above, a risk of failure • 40 degree slope and above has a 50% chance of failure • 45 degree slope and above 85% chance of failure